Fix speed

+++ beholder-cerebrum/data/node-red-contrib-beholder/nodes/stalk/move-mouse-to/move-mouse-to.js
@@ -28,7 +28,7 @@ module.exports = function (RED) {
           current_position: currentPosition,
           target_position: targetPosition,
           movement_type: 0,
-          movement_speed: parseFloat(config.movementSpeed),
+          movement_speed: parseInt(config.movementSpeed),
           pre_move_actions: config.preMoveActions,
           post_move_actions: config.postMoveActions,
         }
